+++ examples/rand/rand.go
@@ -7,7 +7,6 @@ import (
 	"runtime"
 
 	"github.com/open-quantum-safe/liboqs-go/oqs"
-	oqsrand "github.com/open-quantum-safe/liboqs-go/oqs/rand" // RNG support
 )
 
 // CustomRNG provides a (trivial) custom random number generator; the memory is
@@ -22,20 +21,20 @@ func CustomRNG(randomArray []byte, bytesToRead int) {
 func main() {
 	fmt.Println("liboqs version: " + oqs.LiboqsVersion())
 
-	if err := oqsrand.RandomBytesSwitchAlgorithm("system"); err != nil {
+	if err := oqs.RandomBytesSwitchAlgorithm("system"); err != nil {
 		log.Fatal(err)
 	}
-	fmt.Printf("%18s% X\n", "System (default): ", oqsrand.RandomBytes(32))
-	if err := oqsrand.RandomBytesCustomAlgorithm(CustomRNG); err != nil {
+	fmt.Printf("%18s% X\n", "System (default): ", oqs.RandomBytes(32))
+	if err := oqs.RandomBytesCustomAlgorithm(CustomRNG); err != nil {
 		log.Fatal(err)
 	}
-	fmt.Printf("%-18s% X\n", "Custom RNG: ", oqsrand.RandomBytes(32))
+	fmt.Printf("%-18s% X\n", "Custom RNG: ", oqs.RandomBytes(32))
 
 	// We do not yet support OpenSSL under Windows
 	if runtime.GOOS != "windows" {
-		if err := oqsrand.RandomBytesSwitchAlgorithm("OpenSSL"); err != nil {
+		if err := oqs.RandomBytesSwitchAlgorithm("OpenSSL"); err != nil {
 			log.Fatal(err)
 		}
-		fmt.Printf("%-18s% X\n", "OpenSSL: ", oqsrand.RandomBytes(32))
+		fmt.Printf("%-18s% X\n", "OpenSSL: ", oqs.RandomBytes(32))
 	}
 }